+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to both [Semantic Versioning](https://semver.org/spec/v2.0.
 - This CHANGELOG file.
 - pre-commit config.
 - test coverage using Codecov
+- LimitedTaskGroup to execute tasks limiting concurrency
 
 ### Changed